+++ tesseract_rviz/src/ros_environment_widget.cpp
@@ -1,5 +1,5 @@
-#include <tesseract_rviz/environment_plugin/ros_environment_widget.h>
-#include <tesseract_rviz/environment_plugin/conversions.h>
+#include <tesseract_rviz/ros_environment_widget.h>
+#include <tesseract_rviz/conversions.h>
 #include <tesseract_rviz/conversions.h>
 
 #include <tesseract_rviz/markers/marker_base.h>
@@ -249,8 +249,7 @@ void ROSEnvironmentWidget::onRender(float dt)
                 break;
               }
               case tesseract_environment::CommandType::CHANGE_LINK_COLLISION_ENABLED:
-              case tesseract_environment::CommandType::ADD_ALLOWED_COLLISION:
-              case tesseract_environment::CommandType::REMOVE_ALLOWED_COLLISION:
+              case tesseract_environment::CommandType::MODIFY_ALLOWED_COLLISIONS:
               case tesseract_environment::CommandType::REMOVE_ALLOWED_COLLISION_LINK:
               case tesseract_environment::CommandType::CHANGE_JOINT_POSITION_LIMITS:
               case tesseract_environment::CommandType::CHANGE_JOINT_VELOCITY_LIMITS:
@@ -311,7 +310,7 @@ void ROSEnvironmentWidget::onRender(float dt)
             auto container = data_->entity_manager->getEntityContainer(pair.first);
             Ogre::Vector3 position;
             Ogre::Quaternion orientation;
-            toOgre(position, orientation, pair.second);
+            toOgreLegacy(position, orientation, pair.second);
 
             auto entity = container->getTrackedEntity(tesseract_gui::EntityContainer::VISUAL_NS, pair.first);
             Ogre::SceneNode* sn = data_->scene_manager->getSceneNode(entity.unique_name);